rerouting all the info to a separate file for classes
